move role validator out of core/config to modelhub layer
The validator needs the installed registry and the catalog to know each
ref's task, but core/config is the foundation layer (Layer 0) and is
not allowed to import from modelhub or providers. The previous file
already cheated with PROVIDER_PREFIXES at module top and lazy imports
of lilbee.catalog; the fix in this branch added two more lazy imports
of modelhub, increasing the upward-import count flagged by the
AGENTS.md code-smell sweep.

Slim core/config/validators.py back to just ConfigField (the pydantic
helper that genuinely belongs at Layer 0). Move TaskMismatchError,
validate_model_task_assignment, and the small helpers around them to
the new src/lilbee/modelhub/role_validator.py, which can import the
registry, the discovery reclassifier, the catalog, and PROVIDER_PREFIXES
directly. Every consumer (TUI, HTTP handlers, HTTP routes, tests)
updates to the new path in the same commit. No re-export shim.
