+++ src/lilbee/modelhub/role_validator.py
@@ -1,0 +1,101 @@
+"""Role-slot assignment validation for the four model config fields."""
+
+import os
+import sys
+from typing import Any
+
+from lilbee.catalog import find_catalog_entry
+from lilbee.catalog.types import ModelTask
+from lilbee.core.config import cfg
+from lilbee.modelhub.model_manager.discovery import reclassify_by_name
+from lilbee.modelhub.registry import ModelRegistry
+from lilbee.providers.model_ref import PROVIDER_PREFIXES
+
+# Test-only bypass. Both the env var and pytest must be present so a
+# leaked env var cannot disable validation in production.
+_SKIP_MODEL_TASK_VALIDATION_ENV = "LILBEE_SKIP_MODEL_TASK_VALIDATION"
+
+
+def _model_task_validation_bypassed() -> bool:
+    if not os.environ.get(_SKIP_MODEL_TASK_VALIDATION_ENV):
+        return False
+    return sys.modules.get("pytest") is not None
+
+
+_MODEL_FIELD_TO_TASK: dict[str, str] = {
+    "chat_model": "chat",
+    "embedding_model": "embedding",
+    "vision_model": "vision",
+    "reranker_model": "rerank",
+}
+
+
+class TaskMismatchError(ValueError):
+    """A role slot was assigned a model whose catalog task does not match.
+
+    Carries the structured fields so each surface (HTTP, CLI, TUI, MCP)
+    can format its own user-facing message. The default ``str()`` form is
+    surface-neutral so it is safe to surface unmodified.
+    """
+
+    def __init__(self, ref: str, entry_task: ModelTask, expected_task: ModelTask) -> None:
+        self.ref = ref
+        self.entry_task = entry_task
+        self.expected_task = expected_task
+        super().__init__(f"Model '{ref}' is a {entry_task} model, not {expected_task}.")
+
+
+# A native GGUF ref of the form ``<owner>/<repo>/<file>.gguf`` has at least
+# two ``/`` separators; one-slash refs are bare repo IDs.
+_NATIVE_GGUF_REF_MIN_SLASHES = 2
+
+
+def _resolve_installed_task(ref: str) -> ModelTask | None:
+    """Return the manifest's ``ModelTask`` for *ref*, name-reclassified, or ``None``."""
+    manifest = ModelRegistry(cfg.models_dir).get_manifest(ref)
+    if manifest is None:
+        return None
+    return ModelTask(reclassify_by_name(ref, manifest.task))
+
+
+def _skips_catalog_check(ref: str, *, allow_bypass: bool) -> bool:
+    """Whether *ref* skips the catalog check."""
+    if not ref or not ref.strip():
+        return True
+    if allow_bypass and _model_task_validation_bypassed():
+        return True
+    return ref.split("/", 1)[0] in PROVIDER_PREFIXES
+
+
+def validate_model_task_assignment(field_name: str, ref: str, *, allow_bypass: bool = True) -> str:
+    """Check *ref* is assignable to *field_name*; return the canonical ref.
+
+    Accepts featured catalog refs and installed non-featured refs (any model
+    the user has pulled). Raises ``TaskMismatchError`` on role mismatch and
+    ``ValueError`` when the model is neither featured nor installed.
+    """
+    if _skips_catalog_check(ref, allow_bypass=allow_bypass):
+        return ref
+    want = ModelTask(_MODEL_FIELD_TO_TASK[field_name])
+
+    entry: Any = find_catalog_entry(ref)
+    if entry is not None:
+        if entry.task != want:
+            raise TaskMismatchError(ref, ModelTask(entry.task), want)
+        # Keep a full ``<repo>/<file>.gguf`` so resolve_model_path lands on
+        # the exact installed quant; fall back to the catalog ref otherwise.
+        if ref.endswith(".gguf") and ref.count("/") >= _NATIVE_GGUF_REF_MIN_SLASHES:
+            return ref
+        canonical: str = entry.ref
+        return canonical
+
+    installed_task = _resolve_installed_task(ref)
+    if installed_task is None:
+        raise ValueError(
+            f"Model '{ref}' is not installed. "
+            "Install it with 'lilbee model pull <ref>' "
+            "(or POST /api/models/pull) before assigning it to a role."
+        )
+    if installed_task != want:
+        raise TaskMismatchError(ref, installed_task, want)
+    return ref
