+++ tests/test_server_litestar.py
@@ -11,7 +11,7 @@ from litestar.testing import TestClient
 
 from lilbee.catalog.types import ModelTask
 from lilbee.core.config import cfg
-from lilbee.core.config.validators import TaskMismatchError
+from lilbee.modelhub.role_validator import TaskMismatchError
 
 
 @pytest.fixture(autouse=True)
